Support for transparency in images
The ECMA48 sixel parser now recognizes transparency and honors it.

When using the Swing backend, if jexer.Swing.imagesOverText is true
then text glyphs will be rendered underneath image cells.  This is
defaulted to false because of the hideous performance cost at the
moment, especially against sixel animations in terminal windows.

ECMA48 backend is still unaware of transparency.  We would need to
define a font with fallback and render images-over-text to that, which
is a bit more than I want to handle at the moment.

Jexer PNG images also handle 24-bit + transparency.  One can use the
'imgls' script in Jexer's examples directory to try this out.